LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/snap-after-initial-layout/initial-snap-matches-scroll-expected.txt




FAIL initial scroll snap matches scrolling scroll snap position. assert_equals: expected 351 but got 956
